Parte 10 da aula 1 do curso "Android parte 3: refinando o projeto" em Kotlin.
